Attempt at rewrite of a note on properties
Still not sure if this is accurate: Name and Id of a type are shown on its Properties field. Similarly, Name and Id of a nested type are included in its Properties. If there are two types with the same Properties, Name or ID, the search will return both types in the result.